Datatype of the method return changed so that it can work with Atomic Data type

+++ NationalBankOfMDX/src/nationalbankofmdx/User.java
@@ -50,7 +50,7 @@ public class User {
         }
     }
 
-    public double userBalance() {
+    public AtomicInteger userBalance() {
         return bankAccount.getAccountBalance();
     }     
 }
